+++ config/configmap.go
@@ -30,6 +30,11 @@ const (
 	KeyDelimiter = "::"
 )
 
+// MapConverterFunc is a converter function for the config.Map that allows distributions
+// (in the future components as well) to build backwards compatible config converters.
+// TODO: Consider to add a context as the first argument.
+type MapConverterFunc func(*Map) error
+
 // NewMap creates a new empty config.Map instance.
 func NewMap() *Map {
 	return &Map{k: koanf.New(KeyDelimiter)}